# backend/src/main/java/com/enterprise/knowledge/dto/response/ConversationResponse.java
package com.enterprise.knowledge.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.Instant;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ConversationResponse {

    private UUID id;
    private String title;
    private boolean isPinned;
    private int messageCount;
    private Instant lastMessageAt;
    private Instant createdAt;
}