# docs/api/api-contracts.yaml
openapi: 3.1.0
info:
  title: Enterprise AI Knowledge Assistant API
  description: |
    Production-grade RAG-powered knowledge assistant API.
    Enables semantic search and AI-powered Q&A over enterprise documents.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: platform@enterprise.com
servers:
  - url: https://api.knowledge.enterprise.com/v1
    description: Production
  - url: https://api-staging.knowledge.enterprise.com/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /chat/conversations:
    get:
      tags: [Chat]
      summary: List user conversations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: pinned
          in: query
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationPage'
          description: Conversation list

    post:
      tags: [Chat]
      summary: Start a new conversation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
          description: New conversation
  /chat/conversations/{id}:
    get:
      tags: [Chat]
      summary: Get conversation with message history
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'
          description: Conversation with messages

    delete:
      tags: [Chat]
      summary: Delete a conversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '204':
          description: Deleted
  /chat/conversations/{id}/pin:
    put:
      tags: [Chat]
      summary: Toggle pin status of a conversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
          description: Updated conversation
  /chat/ask:
    post:
      tags: [Chat]
      summary: Ask a question (non-streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
          description: AI answer with citations
  /chat/stream:
    post:
      tags: [Chat]
      summary: Ask a question with Server-Sent Events streaming
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: SSE stream of tokens
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type":"token","content":"The "}
                  data: {"type":"token","content":"authentication "}
                  data: {"type":"citations","citations":[{"index":1,"documentId":"...","title":"Auth Guide"}]}
                  data: {"type":"done","messageId":"...","conversationId":"..."}
  /chat/messages/{id}/regenerate:
    post:
      tags: [Chat]
      summary: Regenerate the AI response for a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
          description: Regenerated response
  /chat/messages/{id}/feedback:
    post:
      tags: [Chat]
      summary: Submit feedback for an AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Feedback recorded

  # ─── ANALYTICS ─────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ConversationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
    SignupRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        department:
          type: string
        jobTitle:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Access token TTL in seconds
        user:
          $ref: '#/components/schemas/UserProfile'
    TokenRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        department:
          type: string
        jobTitle:
          type: string
        avatarUrl:
          type: string
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
        conversationId:
          type: string
          format: uuid
          description: Omit to start a new conversation
        searchMode:
          type: string
          enum: [HYBRID, VECTOR, KEYWORD]
          default: HYBRID
        topK:
          type: integer
          default: 5
          minimum: 1
          maximum: 20
        documentIds:
          type: array
          items:
            type: string
            format: uuid
          description: Restrict retrieval to specific documents
    ChatResponse:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        tokensUsed:
          type: integer
        latencyMs:
          type: integer
        model:
          type: string
        confidenceScore:
          type: number
          format: float
    Citation:
      type: object
      properties:
        index:
          type: integer
        documentId:
          type: string
          format: uuid
        documentTitle:
          type: string
        chunkId:
          type: string
          format: uuid
        excerpt:
          type: string
        pageNumber:
          type: integer
        sectionTitle:
          type: string
        relevanceScore:
          type: number
          format: float
    ConversationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        isPinned:
          type: boolean
        messageCount:
          type: integer
        lastMessageAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ConversationDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ConversationResponse'
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/MessageResponse'
    MessageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [USER, ASSISTANT]
        content:
          type: string
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        tokensUsed:
          type: integer
        latencyMs:
          type: integer
        createdAt:
          type: string
          format: date-time
    ConversationPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ConversationResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
    FeedbackRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          enum: [1, 5]
          description: 1 = thumbs down, 5 = thumbs up
        comment:
          type: string
        feedbackType:
          type: string
          enum: [HELPFUL, INACCURATE, INCOMPLETE, HARMFUL]
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  rejectedValue:
                    type: string